# src/Parbat.Core/Parbat.Core.API/Controllers/CurriculumTypeController.cs
using Microsoft.AspNetCore.Http;
using Microsoft.AspNetCore.Mvc;
using Parbat.Core.DataObjects;
using Parbat.Core.Services;
using System.Collections.Generic;

namespace Parbat.Core.API.Controllers
{
    /// <summary>
    /// CurriculumType
    /// </summary>
    [Route(Global.API_CONTROLLER)]
    [ApiController]
    public class CurriculumTypeController : Controller
    {
        private CurriculumTypeService _service;

        /// <summary>
        /// Constructor 
        /// </summary>
        /// <param name="service">IRepositoryFactory</param>
        public CurriculumTypeController(CurriculumTypeService service)
        {
            _service = service;
        }

        /// <summary>
        /// Get a record
        /// </summary>
        /// <param name="id"></param>
        /// <returns></returns>
        [HttpGet("{id}", Name = "CurriculumTypeGetById")]
        public ActionResult<CurriculumType> Get(long id)
        {
            try
            {
                CurriculumType found = _service.FindByID(id);

                return Ok(found);
            }
            catch (ServiceException ex)
            {
                return BadRequest(ex.Message);
            }
        }

        /// <summary>
        /// List all curriculumType
        /// </summary>
        /// <returns></returns>
        [HttpGet(Name = "CurriculumTypeList")]
        public ActionResult<List<CurriculumType>> List()
        {
            try
            {
                var curriculumTypes = _service.GetAll();

                return Ok(curriculumTypes);
            }
            catch (ServiceException ex)
            {
                return BadRequest(ex.Message);
            }
        }

        /// <summary>
        /// Update curriculumType
        /// </summary>
        /// <param name="c"></param>
        /// <returns></returns>
        [HttpPut(Name = "CurriculumTypeUpdate")]
        public ActionResult<CurriculumType> Update([FromBody] CurriculumType c)
        {
            try
            {
                _service.Update(c);

                return NoContent();
            }
            catch (ServiceException se)
            {
                return BadRequest(se.Message);
            }
        }

        /// <summary>
        /// Create 
        /// </summary>
        /// <param name="c"></param>
        /// <returns></returns>
        [HttpPost(Name = "CurriculumTypeCreate")]
        [ProducesResponseType(typeof(CurriculumType), StatusCodes.Status201Created)]
        public ActionResult<CurriculumType> Create([FromBody] CurriculumType c)
        {
            try
            {
                _service.Create(c);

                return Created("Get", c);
            }
            catch (ServiceException se)
            {
                return BadRequest(se.Message);
            }

        }

        /// <summary>
        /// Delete
        /// </summary>
        /// <param name="id"></param>
        /// <returns></returns>
        [HttpDelete("{id}", Name = "CurriculumTypeDeleteById")]
        [ProducesResponseType(StatusCodes.Status204NoContent)]
        public ActionResult<long> Delete(long id)
        {
            try
            {
                _service.Delete(id);

                return NoContent();
            }
            catch (ServiceException se)
            {
                return BadRequest(se.Message);
            }
        }
    }
}