# src/Parbat.Core/Parbat.Core.Web/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Parbat.Core.API",
    "description": "An open source ERP for Universities",
    "contact": {
      "name": "Dr. Muhammad Qasim Pasta",
      "url": "https://twitter.com/mqpasta",
      "email": "qasim.pasta@ieee.org"
    },
    "license": {
      "name": "The MIT License (MIT)",
      "url": "http://opensource.org/licenses/mit-license.php"
    },
    "version": "v2.0"
  },
  "paths": {
...
    "/api/v2.0/Curriculum/Get/{id}": {
      "get": {
        "tags": [
          "Curriculum"
        ],
        "summary": "Get a record",
        "operationId": "CurriculumGetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Curriculum"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Curriculum"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Curriculum"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2.0/Curriculum/Update": {
      "put": {
        "tags": [
          "Curriculum"
        ],
        "summary": "Update a curriculum",
        "operationId": "CurriculumUpdate",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Curriculum"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Curriculum"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Curriculum"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Curriculum"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Curriculum"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Curriculum"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2.0/Curriculum/Create": {
      "post": {
        "tags": [
          "Curriculum"
        ],
        "summary": "Create",
        "operationId": "CurriculumCreate",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Curriculum"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Curriculum"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Curriculum"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Curriculum"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Curriculum"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Curriculum"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2.0/Curriculum/Delete/{id}": {
      "delete": {
        "tags": [
          "Curriculum"
        ],
        "summary": "Delete",
        "operationId": "CurriculumDeleteById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/v2.0/Curriculum/GetOwnerUnit/OwnerUnit/{id}": {
      "get": {
        "tags": [
          "Curriculum"
        ],
        "summary": "return the curriculum owners",
        "operationId": "CurriculumGetOwnerUnitById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Curriculum"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Curriculum"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Curriculum"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v2.0/Curriculum/GetCurriculumType/CurriculumType/{id}": {
      "get": {
        "tags": [
          "Curriculum"
        ],
        "summary": "Return all curriculumTypes",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Curriculum"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Curriculum"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Curriculum"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v2.0/CurriculumType/Get/{id}": {
      "get": {
        "tags": [
          "CurriculumType"
        ],
        "summary": "Get a record",
        "operationId": "CurriculumTypeGetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CurriculumType"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurriculumType"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurriculumType"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2.0/CurriculumType/Update": {
      "put": {
        "tags": [
          "CurriculumType"
        ],
        "summary": "Update curriculumType",
        "operationId": "CurriculumTypeUpdate",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CurriculumType"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CurriculumType"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CurriculumType"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CurriculumType"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurriculumType"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurriculumType"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2.0/CurriculumType/Create": {
      "post": {
        "tags": [
          "CurriculumType"
        ],
        "summary": "Create",
        "operationId": "CurriculumTypeCreate",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CurriculumType"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CurriculumType"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CurriculumType"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CurriculumType"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurriculumType"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurriculumType"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2.0/CurriculumType/Delete/{id}": {
      "delete": {
        "tags": [
          "CurriculumType"
        ],
        "summary": "Delete",
        "operationId": "CurriculumTypeDeleteById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Curriculum": {
        "required": [
          "CurriculumTypeID",
          "IsLimitCourses",
          "IsLimitCreditHrs",
          "Name",
          "OwnerUnitID",
          "ShortName",
          "TotalCourses",
          "TotalCreditHrs"
        ],
        "type": "object",
        "properties": {
          "CurriculumID": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "Name": {
            "minLength": 1,
            "type": "string"
          },
          "ShortName": {
            "minLength": 1,
            "type": "string"
          },
          "OwnerUnitID": {
            "type": "integer",
            "format": "int64"
          },
          "TotalCourses": {
            "type": "integer",
            "format": "int32"
          },
          "TotalCreditHrs": {
            "type": "integer",
            "format": "int32"
          },
          "IsLimitCourses": {
            "type": "boolean"
          },
          "IsLimitCreditHrs": {
            "type": "boolean"
          },
          "CurriculumTypeID": {
            "type": "integer",
            "format": "int64"
          },
          "Key": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "CurriculumType": {
        "required": [
          "Name"
        ],
        "type": "object",
        "properties": {
          "CurriculumTypeID": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "Name": {
            "minLength": 1,
            "type": "string"
          },
          "Key": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}